requirements.txt
asgiref==3.5.2
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.1.4
django-cors-headers==3.13.0
djangorestframework==3.14.0
idna==3.4
mysql-client==0.0.1
Pillow==9.3.0
python-decouple==3.6
pytz==2022.6
requests==2.28.1
sqlparse==0.4.3
urllib3==1.26.13
